criação de classes bean (lista) e ajuste context XML

+++ src/main/java/br/gov/pr/legislativo/entidades/TipoProposicao.java
@@ -1,29 +1,29 @@
 package br.gov.pr.legislativo.entidades;
 
 import java.io.Serializable;
-import javax.persistence.CascadeType;
+//import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
+//import javax.persistence.JoinColumn;
+//import javax.persistence.ManyToOne;
 
 @Entity
 public class TipoProposicao implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long idTipoProposicao;
+	private Integer idTipoProposicao;
 	@Column(nullable = false)
 	private String nomeTipo;
 
-	public Long getIdTipoProposicao() {
+	public Integer getIdTipoProposicao() {
 		return idTipoProposicao;
 	}
 
-	public void setIdTipoProposicao(Long idTipoProposicao) {
+	public void setIdTipoProposicao(Integer idTipoProposicao) {
 		this.idTipoProposicao = idTipoProposicao;
 	}
